refactor: renamed block_committer/types => block_committer/input
