+++ crates/committer/src/block_committer.rs
@@ -1,1 +1,1 @@
-pub mod types;
+pub mod input;
