+++ crates/committer_cli/src/parse_input/cast.rs
@@ -1,5 +1,5 @@
 use crate::parse_input::types::RawInput;
-use committer::block_committer::types::{
+use committer::block_committer::input::{
     ContractAddress, ContractState, Input, StarknetStorageKey, StarknetStorageValue, StateDiff,
 };
 use committer::felt::Felt;
