+++ crates/committer_cli/src/parse_input/read.rs
@@ -1,4 +1,4 @@
-use committer::{block_committer::types::Input, storage::errors::DeserializationError};
+use committer::{block_committer::input::Input, storage::errors::DeserializationError};
 
 use crate::parse_input::types::RawInput;
